+++ game/exceptions.py
@@ -6,3 +6,12 @@ class InvalidMove(Exception):
 
 class OutOfBoard(Exception):
     pass
+
+class GoingThroughAPiece(Exception):
+    pass
+
+class SameColor(Exception):
+    pass
+
+class WrongPiece(Exception):
+    pass